+++ crates/ruma-client-api/src/lib.rs
@@ -12,7 +12,6 @@
 pub mod account;
 pub mod alias;
 pub mod appservice;
-#[cfg(feature = "unstable-msc3916")]
 pub mod authenticated_media;
 pub mod backup;
 pub mod config;
@@ -24,6 +23,8 @@ pub mod directory;
 pub mod discovery;
 pub mod error;
 pub mod filter;
+#[cfg(feature = "unstable-msc4140")]
+pub mod future;
 pub mod http_headers;
 pub mod keys;
 pub mod knock;