+++ crates/ruma-events/src/enums.rs
@@ -45,9 +45,8 @@ event_enum! {
         "m.call.candidates" => super::call::candidates,
         "m.call.negotiate" => super::call::negotiate,
         "m.call.reject" => super::call::reject,
-        #[cfg(feature = "unstable-msc3291")]
-        #[ruma_enum(alias = "m.call.sdp_stream_metadata_changed")]
-        "org.matrix.call.sdp_stream_metadata_changed" => super::call::sdp_stream_metadata_changed,
+        #[ruma_enum(alias = "org.matrix.call.sdp_stream_metadata_changed")]
+        "m.call.sdp_stream_metadata_changed" => super::call::sdp_stream_metadata_changed,
         "m.call.select_answer" => super::call::select_answer,
         #[cfg(feature = "unstable-msc3954")]
         #[ruma_enum(alias = "m.emote")]
@@ -88,6 +87,9 @@ event_enum! {
         #[cfg(feature = "unstable-msc3381")]
         #[ruma_enum(ident = UnstablePollEnd)]
         "org.matrix.msc3381.poll.end" => super::poll::unstable_end,
+        #[cfg(feature = "unstable-msc3489")]
+        #[ruma_enum(alias = "m.beacon")]
+        "org.matrix.msc3672.beacon" => super::beacon,
         "m.reaction" => super::reaction,
         "m.room.encrypted" => super::room::encrypted,
         "m.room.message" => super::room::message,
@@ -127,6 +129,9 @@ event_enum! {
         "m.room.topic" => super::room::topic,
         "m.space.child" => super::space::child,
         "m.space.parent" => super::space::parent,
+        #[cfg(feature = "unstable-msc3489")]
+        #[ruma_enum(alias = "m.beacon_info")]
+        "org.matrix.msc3672.beacon_info" => super::beacon_info,
         #[cfg(feature = "unstable-msc3401")]
         #[ruma_enum(alias = "m.call.member")]
         "org.matrix.msc3401.call.member" => super::call::member,
@@ -309,6 +314,8 @@ impl AnyMessageLikeEventContent {
     /// This is a helper function intended for encryption. There should not be a reason to access
     /// `m.relates_to` without first destructuring an `AnyMessageLikeEventContent` otherwise.
     pub fn relation(&self) -> Option<encrypted::Relation> {
+        #[cfg(feature = "unstable-msc3489")]
+        use super::beacon::BeaconEventContent;
         use super::key::verification::{
             accept::KeyVerificationAcceptEventContent, cancel::KeyVerificationCancelEventContent,
             done::KeyVerificationDoneEventContent, key::KeyVerificationKeyEventContent,
@@ -361,13 +368,16 @@ impl AnyMessageLikeEventContent {
             | Self::UnstablePollEnd(UnstablePollEndEventContent { relates_to, .. }) => {
                 Some(encrypted::Relation::Reference(relates_to.clone()))
             }
+            #[cfg(feature = "unstable-msc3489")]
+            Self::Beacon(BeaconEventContent { relates_to, .. }) => {
+                Some(encrypted::Relation::Reference(relates_to.clone()))
+            }
             #[cfg(feature = "unstable-msc3381")]
             Self::PollStart(_) | Self::UnstablePollStart(_) => None,
             #[cfg(feature = "unstable-msc4075")]
             Self::CallNotify(_) => None,
-            #[cfg(feature = "unstable-msc3291")]
-            Self::CallSdpStreamMetadataChanged(_) => None,
-            Self::CallNegotiate(_)
+            Self::CallSdpStreamMetadataChanged(_)
+            | Self::CallNegotiate(_)
             | Self::CallReject(_)
             | Self::CallSelectAnswer(_)
             | Self::CallAnswer(_)